rephrased activity submission notification message. moved magenic dataservice to Common project.
