+++ Magenic.BadgeApplication/Magenic.BadgeApplication.Common/Utils/RestClientFactory.cs
@@ -2,10 +2,18 @@
 using RestSharp;
 using System;
 
-namespace Magenic.BadgeApplication.Teams.Utilities
+namespace Magenic.BadgeApplication.Common.Utilities
 {
+    /// <summary>
+    /// 
+    /// </summary>
     public class RestClientFactory : IRestClientFactory
     {
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="url"></param>
+        /// <returns></returns>
         public IRestClient Create(Uri url)
         {
             return new RestClient(url);
